added popup menu on scene graph nodes that allows you to delete a node
made Scene objects extend BaseSceneNode and now the current scene is the root of the scene graph

fixed bug with PanelRenderer where if an animation was present, it was always using the first frame of the animation, regardless of what the startframe was set to. Now, a PanelRenderer with a KeyFrameAnimation attached will render the start frame of the animation.